pyproject.toml
[project]
name = "booking-automation"
version = "0.1.0"
description = "Automate Microsoft Bookings for GTA Office Hours scheduling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.14.3",
]

[project.scripts]
book = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]
